Moved rest of functions so far into proper headers
