+++ VentilatorControl/src/states.h
@@ -1,0 +1,7 @@
+#ifndef __STATES_H__
+#define __STATES_H__
+
+void cycleStateHandler (void);
+void cycleStateSetpointHandler(void);
+
+#endif // __STATES_H__
